:new: [gherkin] Move `gherkin` support to core
Problem:
- Gherkin is only supported in the example.

Solution:
- Move Gherkin support to the core.
