+++ include/boost/ut.hpp
@@ -196,6 +196,43 @@ class function<R(TArgs...)> {
   return is_match(input.substr(1), pattern.substr(1));
 }
 
+template <class TPattern, class TStr>
+[[nodiscard]] constexpr auto match(const TPattern& pattern, const TStr& str)
+    -> std::vector<TStr> {
+  std::vector<TStr> groups{};
+  auto pi = 0u, si = 0u;
+
+  const auto matcher = [&](char b, char e, char c = 0) {
+    const auto match = si;
+    while (str[si] and str[si] != b and str[si] != c) {
+      ++si;
+    }
+    groups.emplace_back(str.substr(match, si - match));
+    while (pattern[pi] and pattern[pi] != e) {
+      ++pi;
+    }
+    pi++;
+  };
+
+  while (pi < std::size(pattern) && si < std::size(str)) {
+    if (pattern[pi] == '\'' and str[si] == '\'' and pattern[pi + 1] == '{') {
+      ++si;
+      matcher('\'', '}');
+    } else if (pattern[pi] == '{') {
+      matcher(' ', '}', ',');
+    } else if (pattern[pi] != str[si]) {
+      return {};
+    }
+    ++pi, ++si;
+  }
+
+  if (si < str.size() or pi < std::size(pattern)) {
+    return {};
+  }
+
+  return groups;
+}
+
 template <class T = std::string_view, class TDelim>
 [[nodiscard]] inline auto split(T input, TDelim delim) -> std::vector<T> {
   std::vector<T> output{};
@@ -2291,6 +2328,120 @@ namespace bdd {
 [[maybe_unused]] constexpr auto then = [](const auto name) {
   return detail::test{"then", name};
 };
+#if not defined(BOOST_UT_FORWARD)
+namespace gherkin {
+class steps {
+  using step_t = std::string;
+  using steps_t = void (*)(steps&);
+  using gherkin_t = std::vector<step_t>;
+  using call_step_t = utility::function<void(const std::string&)>;
+  using call_steps_t = std::vector<std::pair<step_t, call_step_t>>;
+
+  class step {
+   public:
+    template <class TPattern>
+    step(steps& steps, const TPattern& pattern)
+        : steps_{steps}, pattern_{pattern} {}
+
+    ~step() { steps_.next(pattern_); }
+
+    template <class TExpr>
+    auto operator=(const TExpr& expr) -> void {
+      for (const auto& [pattern, _] : steps_.call_steps()) {
+        if (pattern_ == pattern) {
+          return;
+        }
+      }
+
+      steps_.call_steps().emplace_back(
+          pattern_, [expr, pattern = pattern_](const auto& step) {
+            [=]<class... TArgs>(type_traits::list<TArgs...>) {
+              log << step;
+              auto i = 0;
+              const auto& ms = utility::match(pattern, step);
+              expr(TArgs{std::stoi(ms[i++])}...);
+            }
+            (typename type_traits::function_traits<TExpr>::args{});
+          });
+    }
+
+   private:
+    steps& steps_;
+    std::string pattern_{};
+  };
+
+ public:
+  template <class TSteps>
+  constexpr /*explicit(false)*/ steps(const TSteps& steps) : steps_{steps} {}
+
+  template <class TGherkin>
+  auto operator|(const TGherkin& gherkin) {
+    gherkin_ = utility::split<std::string>(gherkin, '\n');
+    for (auto& step : gherkin_) {
+      step.erase(0, step.find_first_not_of(" \t"));
+    }
+
+    return [this] {
+      step_ = {};
+      steps_(*this);
+    };
+  }
+  auto feature(const std::string& pattern) {
+    return step{*this, "Feature: " + pattern};
+  }
+  auto scenario(const std::string& pattern) {
+    return step{*this, "Scenario: " + pattern};
+  }
+  auto given(const std::string& pattern) {
+    return step{*this, "Given " + pattern};
+  }
+  auto when(const std::string& pattern) {
+    return step{*this, "When " + pattern};
+  }
+  auto then(const std::string& pattern) {
+    return step{*this, "Then " + pattern};
+  }
+
+ private:
+  template <class TPattern>
+  auto next(const TPattern& pattern) -> void {
+    const auto is_scenario = [&pattern](const auto& step) {
+      constexpr auto scenario = "Scenario";
+      return pattern.find(scenario) == std::string::npos and
+             step.find(scenario) != std::string::npos;
+    };
+
+    const auto call_steps = [this, is_scenario](const auto& step,
+                                                const auto i) {
+      for (const auto& [name, call] : call_steps_) {
+        if (is_scenario(step)) {
+          break;
+        }
+
+        if (utility::is_match(step, name) or
+            not std::empty(utility::match(name, step))) {
+          step_ = i;
+          call(step);
+        }
+      }
+    };
+
+    for (auto i = 0; const auto& step : gherkin_) {
+      if (i++ == step_) {
+        call_steps(step, i);
+      }
+    }
+  }
+
+  auto call_steps() -> call_steps_t& { return call_steps_; }
+
+  steps_t steps_{};
+  gherkin_t gherkin_{};
+  call_steps_t call_steps_{};
+  decltype(sizeof("")) step_{};
+};
+}  // namespace gherkin
+#endif
 }  // namespace bdd
 
 namespace spec {
